requirements.txt
streamlit
langdetect
requests
faiss-cpu
sentence-transformers
huggingface_hub
python-dotenv
